refactor: move all ID type definitions to identifiers module
also, add NULL definitions for all cell types
