+++ honeycomb-core/src/cells/identifiers.rs
@@ -5,3 +5,38 @@
 //! Content description if needed
 
 // ------ CONTENT
+
+/// Type definition for dart identifiers
+///
+/// This is used for better control over memory usage and ID encoding.
+pub type DartIdentifier = u32;
+
+pub const NULL_DART_ID: DartIdentifier = 0;
+
+/// Type definition for vertex identifiers
+///
+/// This is used for better control over memory usage and ID encoding.
+pub type VertexIdentifier = u32;
+
+pub const NULL_VERTEX_ID: DartIdentifier = 0;
+
+/// Type definition for edge identifiers
+///
+/// This is used for better control over memory usage and ID encoding.
+pub type EdgeIdentifier = u32;
+
+pub const NULL_EDGE_ID: DartIdentifier = 0;
+
+/// Type definition for face identifiers
+///
+/// This is used for better control over memory usage and ID encoding.
+pub type FaceIdentifier = u32;
+
+pub const NULL_FACE_ID: DartIdentifier = 0;
+
+/// Type definition for volume identifiers
+///
+/// This is used for better control over memory usage and ID encoding.
+pub type VolumeIdentifier = u32;
+
+pub const NULL_VOLUME_ID: DartIdentifier = 0;
